% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{get_current_providers}
\alias{get_current_providers}
\title{Fetch leaflet providers from Leaflet.js.}
\usage{
get_current_providers()
}
\value{
List of \code{providers}, \code{providers.details}
}
\description{
Fetch leaflet providers from Leaflet.js.
}
\examples{
get_current_providers()

}
