% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.providers-package.R
\docType{package}
\name{leaflet.providers-package}
\alias{leaflet.providers}
\alias{leaflet.providers-package}
\title{leaflet.providers: Leaflet Providers}
\description{
Contains third-party map tile provider information from 'Leaflet.js', \url{https://github.com/leaflet-extras/leaflet-providers}, to be used with the 'leaflet' R package. Additionally, 'leaflet.providers' enables users to retrieve up-to-date provider information between package updates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/leaflet.providers/}
  \item \url{https://github.com/rstudio/leaflet.providers}
  \item Report bugs at \url{https://github.com/rstudio/leaflet.providers/issues}
}

}
\author{
\strong{Maintainer}: Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID}) [contributor]

Authors:
\itemize{
  \item Leslie Huang \email{lesliehuang@nyu.edu}
}

Other contributors:
\itemize{
  \item Leaflet Providers contributors (Leaflet Providers plugin) [contributor, copyright holder]
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
