% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.details.R
\docType{data}
\name{providers.details}
\alias{providers.details}
\title{Leaflet Providers' details.}
\format{A nested list with 27 elements. Each element is a list containing
\describe{
\item{url}{source URL}
\item{options}{options}
\item{variants}{name variants, if any}
}}
\source{
\url{https://unpkg.com/leaflet-providers}
}
\usage{
providers.details
}
\description{
A dataset containing Leaflet Providers details.
}
\keyword{datasets}
