% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{get_current_version_num}
\alias{get_current_version_num}
\title{Helper function that returns the current version number of Leaflet.js}
\usage{
get_current_version_num(js_path)
}
\arguments{
\item{js_path}{Path to .js URL from which package info will be extracted.}
}
\value{
Current version number.
}
\description{
Helper function that returns the current version number of Leaflet.js
}
