% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{get_providers}
\alias{get_providers}
\title{Fetch leaflet providers from Leaflet.js.}
\usage{
get_providers(version_num = NULL)
}
\value{
List of \code{version number}, \code{providers}, \code{providers.details}
}
\description{
Fetch leaflet providers from Leaflet.js.
}
\examples{
get_providers()
get_providers("1.8.0")

}
