% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{get_providers}
\alias{get_providers}
\title{Fetch leaflet providers from Leaflet.js.}
\usage{
get_providers(version_num = NULL)
}
\arguments{
\item{version_num}{Version number with which to update leaflet providers. If \code{NULL}, fetches most recent version.}
}
\value{
\code{leaflet_providers} object containing \code{providers_version_num}, \code{providers_data}, \code{providers_details_data}, \code{src}
}
\description{
Fetch leaflet providers from Leaflet.js.
}
\examples{
\donttest{
if (require("V8") && require("jsonlite")) {
  get_providers()
  get_providers("1.8.0")
}
}
}
