% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\docType{data}
\name{providers}
\alias{providers}
\title{Leaflet Providers.}
\format{A list with 155 elements.}
\source{
\url{https://unpkg.com/leaflet-providers}
}
\usage{
providers
}
\description{
A dataset containing the names of leaflet providers.
}
\keyword{datasets}
