% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{providers}
\alias{providers}
\title{Return providers, providers_details, version, and HTML Dependency.}
\usage{
providers()
}
\value{
List of \code{providers_version_num}, \code{providers_data}, \code{providers_details_data}, \code{html_dependency}
}
\description{
Return providers, providers_details, version, and HTML Dependency.
}
\examples{
providers()
}
