% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{providers_default}
\alias{providers_default}
\title{Return default providers, providers_details, version, and HTML Dependency.}
\usage{
providers_default()
}
\value{
\code{leaflet_providers} object containing \code{providers_version_num}, \code{providers},
\code{providers_details}, and \code{src}
}
\description{
Return default providers, providers_details, version, and HTML Dependency.
}
\examples{
str(providers_default(), max = 3, list.len = 4)

}
