% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_providers.R
\name{use_providers}
\alias{use_providers}
\title{Use custom providers data.
Use a custom \code{leaflet_providers} object, e.g. providers data fetched with \link{get_providers}, with the \code{leaflet} package.}
\usage{
use_providers(custom_providers)
}
\arguments{
\item{custom_providers}{A custom \code{leaflet_providers} object.}
}
\description{
Use custom providers data.
Use a custom \code{leaflet_providers} object, e.g. providers data fetched with \link{get_providers}, with the \code{leaflet} package.
}
